New: Add `schema-validation` and `json-parser`  to utils
It was in `hint` where it does not make too much sense. While `hint`
will continue to `export` the types and functions so it is not a
breaking, some references had to be updated because of bundling among
other things.
