+++ packages/hint/package.json
@@ -22,7 +22,6 @@
     "globby": "^9.2.0",
     "is-ci": "^2.0.0",
     "is-svg": "^4.1.0",
-    "jsonc-parser": "^2.1.0",
     "lodash": "^4.17.11",
     "optionator": "^0.8.2",
     "ora": "^3.4.0",
