+++ packages/hint/src/lib/config/config-hints.ts
@@ -10,9 +10,9 @@
 
 import { debug as d } from '@hint/utils/dist/src/debug';
 
-import { validate as schemaValidator } from '../utils/schema-validator';
+import { validate as schemaValidator } from '@hint/utils/dist/src/schema-validation/schema-validator';
 import { HintMetadata, HintConfig } from '../types';
-import { Severity } from '../types/problems';
+import { Severity } from '@hint/utils/dist/src/types/problems';
 
 const debug = d(__filename);
