+++ packages/hint/src/lib/enums/category.ts
@@ -1,11 +1,1 @@
-export enum Category {
-    accessibility = 'accessibility',
-    development = 'development',
-    compatibility = 'compatibility',
-    interoperability = 'interoperability', // Deprecated, to be removed in the next major release
-    other = 'other',
-    pwa = 'pwa',
-    performance = 'performance',
-    pitfalls = 'pitfalls',
-    security = 'security'
-}
+export * from '@hint/utils/dist/src/types/category';
