+++ packages/hint/src/lib/enums/index.ts
@@ -1,4 +1,4 @@
-export * from './category';
+export * from '@hint/utils/dist/src/types/category';
 export * from './error-status';
 export * from './hint-scope';
 export * from './resource-type';
