+++ packages/hint/src/lib/types/analyzer.ts
@@ -1,4 +1,4 @@
-import { Problem } from './problems';
+import { Problem } from '@hint/utils/dist/src/types/problems';
 
 export type CreateAnalyzerOptions = {
     watch?: boolean;
