+++ packages/utils/src/types/category.ts
@@ -1,0 +1,11 @@
+export enum Category {
+    accessibility = 'accessibility',
+    development = 'development',
+    compatibility = 'compatibility',
+    interoperability = 'interoperability', // Deprecated, to be removed in the next major release
+    other = 'other',
+    pwa = 'pwa',
+    performance = 'performance',
+    pitfalls = 'pitfalls',
+    security = 'security'
+}
